requirements.txt
asyncpg
dateparser
discord.py
humanize
jishaku
parsedatetime
psutil
git+https://github.com/Rapptz/discord-ext-menus
